transform html content synchronously
